Add pipes and data validation

+++ src/products/products.controller.ts
@@ -1,6 +1,17 @@
-import { Body, Controller, Delete, Get, Param, Post } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  NotFoundException,
+  Param,
+  ParseUUIDPipe,
+  Post,
+  Put,
+} from '@nestjs/common';
 import { CreateProductDTO } from './dtos/create-product.dto';
 import { ProductsService } from './products.service';
+import { UpdateProductDTO } from './dtos/update-product.dto';
 
 @Controller('products')
 export class ProductsController {
@@ -12,17 +23,37 @@ export class ProductsController {
   }
 
   @Get('/:id')
-  public getById(@Param('id') id: string) {
-    return this.productsService.getById(id);
+  public getById(@Param('id', new ParseUUIDPipe()) id: string) {
+    const prod = this.productsService.getById(id);
+    if (!prod) {
+      throw new NotFoundException('Product not found');
+    }
+    return prod;
   }
 
   @Delete('/:id')
-  public deleteById(@Param('id') id: string) {
+  public deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!this.productsService.getById(id)) {
+      throw new NotFoundException('Product not found');
+    }
     this.productsService.deleteById(id);
     return { success: true };
   }
+
   @Post('/')
-  create(@Body() productData: CreateProductDTO) {
+  public create(@Body() productData: CreateProductDTO) {
     return this.productsService.create(productData);
+  }
+
+  @Put('/:id')
+  public updateById(
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() productData: UpdateProductDTO,
+  ) {
+    if (!this.productsService.getById(id)) {
+      throw new NotFoundException('Product not found');
+    }
+    this.productsService.updateById(id, productData);
+    return { success: true };
   }
 }
